Rename xcomm file to xpython_kernel
